arraysearch.py
A small module which gives advanced search results on searching an item in a list. You get a returned list arranged in descending order of matches, and you can see how much does each element of the list match with searched item. First element is always the best match.
Might be useful for giving the feature of searching items in your game, such as searching weapons or items to craft.